feat(payment): add paypal external service in payment and subscription bounded context.

+++ StockSip.Platform.API/InventoryManagement/Domain/Model/Aggregates/Product.cs
@@ -148,7 +148,7 @@ public partial class Product
     /// <exception cref="ArgumentException">
     /// Thrown when the updated price is less than or equal to zero.
     /// </exception>
-    public void UpdateInformation(string name, string brand, string liquorType, double updatedPrice, int updatedMinimumStock, string updatedImageUrl)
+    public void UpdateInformation(string name, string brand, string liquorType, decimal updatedPrice, int updatedMinimumStock, string updatedImageUrl)
     {
         ProductName = new ProductName(name);
         LiquorType = Enum.Parse<ELiquorType>(liquorType, true);
